% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdprompt.R
\name{init_wd}
\alias{init_wd}
\title{init_wd}
\usage{
init_wd(enabled = TRUE, fullPath = TRUE, promptLen = 15)
}
\arguments{
\item{enabled}{}

\item{fullPath}{}

\item{promptLen}{}
}
\value{
temp
}
\description{
init_wd
}

