% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdprompt.R
\name{start_wd}
\alias{start_wd}
\title{Starts the new behavior for the console prompt.}
\usage{
start_wd()
}
\value{
TRUE if prompt started. FALSE otherwise.
}
\description{
If it is not already running for an interactive session.
}
