% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdprompt.R
\name{wd_prompt}
\alias{wd_prompt}
\title{The real prompt function.}
\usage{
wd_prompt(...)
}
\arguments{
\item{...}{A taskCallback function can be called with 4 or 5 parameters. If
\code{data} is specified, it is always the 5th argument.  Seemed pointless to
declare explicit parameters that would never be used.}
}
\value{
TRUE to continue the taskCallback.  FALSE will delete the taskCallback.
}
\description{
The real prompt function.
}
\seealso{
\link{addTaskCallback}
}
