% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdprompt.R
\name{start_wd}
\alias{start_wd}
\title{Starts the new behavior for the console prompt if it is not already
running for an interactive session.}
\usage{
start_wd()
}
\value{
TRUE if prompt started. FALSE otherwise.
}
\description{
Starts the new behavior for the console prompt if it is not already
running for an interactive session.
}

